+++ tests/test_validation.py
@@ -50,7 +50,7 @@ def test_hydra_partial_via_hydrated_dataclass_is_error():
             pass
 
 
-@given(hydra_partial=st.booleans(), zen_partial=st.booleans())
+@given(hydra_partial=st.booleans(), zen_partial=st.none() | st.booleans())
 def test_specifying_hydra_partial_and_zen_partial_raises(
     hydra_partial: bool, zen_partial: bool
 ):
@@ -153,7 +153,7 @@ def passthrough(*args, **kwargs):
         ),
     ],
 )
-@given(partial=st.booleans(), full_sig=st.booleans())
+@given(partial=st.none() | st.booleans(), full_sig=st.booleans())
 def test_builds_raises_when_user_specified_args_violate_sig(
     func, args, kwargs, full_sig, partial
 ):
@@ -211,7 +211,7 @@ def f(y):
     return y
 
 
-@given(partial=st.booleans(), full_sig=st.booleans())
+@given(partial=st.none() | st.booleans(), full_sig=st.booleans())
 def test_builds_raises_when_base_has_invalid_arg(full_sig, partial):
 
     with pytest.raises(TypeError):
@@ -263,30 +263,6 @@ def f2():
     pass
 
 
-@given(partial=st.booleans(), full_sig=st.booleans())
-def test_builds_raises_when_base_with_partial_target_is_specified(
-    partial: bool, full_sig: bool
-):
-
-    partiald_conf = builds(f2, zen_partial=True)
-
-    if not partial:
-        with pytest.raises(TypeError):
-            builds(
-                f2,
-                populate_full_signature=full_sig,
-                zen_partial=partial,
-                builds_bases=(partiald_conf,),
-            )
-    else:
-        builds(
-            f2,
-            populate_full_signature=full_sig,
-            zen_partial=partial,
-            builds_bases=(partiald_conf,),
-        )
-
-
 def func_with_var_kwargs(**x):
     return x
 
@@ -304,7 +280,7 @@ class Class:
 
 
 @pytest.mark.parametrize("not_callable", [1, "a", None, [1, 2], Class()])
-@given(partial=st.booleans(), full_sig=st.booleans())
+@given(partial=st.none() | st.booleans(), full_sig=st.booleans())
 def test_builds_raises_on_non_callable_target(not_callable, partial, full_sig):
     with pytest.raises(TypeError):
         builds(not_callable, populate_full_signature=full_sig, zen_partial=partial)
@@ -334,7 +310,7 @@ def test_builds_input_validation(param_name: str, value):
 
 def test_just_raises_with_legible_message():
     with pytest.raises(HydraZenUnsupportedPrimitiveError):
-        just(Class())  # type: ignore
+        just(Class())
 
 
 def test_hydrated_dataclass_from_instance_raise():
@@ -347,7 +323,7 @@ def test_hydrated_dataclass_from_instance_raise():
         hydrated_dataclass(dict)(instance_of_a)  # type: ignore
 
 
-@given(partial=st.booleans(), full_sig=st.booleans())
+@given(partial=st.none() | st.booleans(), full_sig=st.booleans())
 def test_builds_raises_for_unimportable_target(partial, full_sig):
     def unreachable():
         pass
@@ -400,7 +376,7 @@ def f_meta_sig(x, *args, y, **kwargs):
         st.sampled_from(["x", "y", "args", "kwargs", "z"]), st.integers()
     ),
     pop_sig=st.booleans(),
-    partial=st.booleans(),
+    partial=st.none() | st.booleans(),
 )
 def test_meta_fields_colliding_with_sig_raises(
     meta_fields, pop_sig: bool, partial: bool
@@ -426,7 +402,7 @@ def test_meta_fields_colliding_with_sig_raises(
     meta_fields=st.dictionaries(
         st.sampled_from(["x", "y", "args", "kwargs", "z"]), st.integers()
     ),
-    partial=st.booleans(),
+    partial=st.none() | st.booleans(),
 )
 def test_meta_fields_colliding_with_user_provided_kwargs_raises(
     meta_fields, partial: bool